breaking down code
