+++ pyrsia_cli/src/cli.rs
@@ -1,0 +1,4 @@
+pub mod handlers;
+pub mod parser;
+
+use crate::commands::*;
